SQL-2167: Organize schema-builder-library and modify function access (#750)
* reorg

* cleanup imports

* limit func visibility

* Update schema-builder-library/src/partitioning/mod.rs


* Update schema-builder-library/src/partitioning/mod.rs


* Update schema-builder-library/src/partitioning/mod.rs


* pr feedback

* turn conversion utils into TryFrom

* cahnge try_into to try_from

---------
