+++ schema-builder-library/src/errors.rs
@@ -16,6 +16,8 @@ pub enum Error {
     NoIdInSample,
     #[error("Driver Error {0}")]
     DriverError(mongodb::error::Error),
+    #[error("Schema Error {0}")]
+    SchemaError(mongosql::schema::Error),
     #[error("NoCollection {0}")]
     NoCollection(String),
     #[error("Execution Error {0}")]
@@ -25,5 +27,15 @@ pub enum Error {
 impl From<mongodb::error::Error> for Error {
     fn from(value: mongodb::error::Error) -> Self {
         Self::DriverError(value)
+    }
+}
+
+impl From<mongosql::schema::Error> for Error {
+    fn from(value: mongosql::schema::Error) -> Self {
+        match value {
+            mongosql::schema::Error::BsonFailure => Self::BsonFailure,
+            mongosql::schema::Error::JsonSchemaFailure => Self::JsonSchemaFailure,
+            _ => Self::SchemaError(value),
+        }
     }
 }
